Post view from mad done, put post in a new js file so it could be used in multiple places
