+++ Views/Navigators/ProgressViewNavigation.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
 import ProgressPresenter from "../../Presenters/ProgressPresenter";
+import Post from "../Post";
 
 const Stack = createStackNavigator();
 
@@ -8,6 +9,7 @@ const ProgressViewNavigator = () => {
     return (
         <Stack.Navigator screenOptions={{headerShown: true}}>
             <Stack.Screen name="Progress" component={ProgressPresenter}/>
+            <Stack.Screen name="Collectible" component={Post}/>
         </Stack.Navigator>
     );
 }
